tests changed

+++ server/src/controllers/users/addressDirection.controller.ts
@@ -4,17 +4,16 @@ import {
   handleError,
   handleObjectNotFound,
 } from "../../utils/error.utils";
-import { extractAuthUserId } from "../../utils/auth.utils";
 import { User } from "../../models/user.model";
 
 class AddressDirectionController {
   public async createAddressDirection(req: Request, res: Response) {
     try {
-      const authUserId = extractAuthUserId(req);
-      const user = await User.findById(authUserId);
+      const { userId } = req.params;
+      const user = await User.findById(userId).select("addressDirections");
       if (!user) return handleObjectNotFound(res, "User");
 
-      user.addressDirections.push({ ...req.body, author: authUserId });
+      user.addressDirections.push(req.body);
 
       const userSaved = await user.save();
       if (!userSaved) return handleBadSaved(res);
@@ -25,14 +24,12 @@ class AddressDirectionController {
   }
   public async updateAddressDirection(req: Request, res: Response) {
     try {
-      const { addressDirectionId } = req.params;
-      const authUserId = extractAuthUserId(req);
-
+      const { userId, addressDirectionId } = req.params;
       const addressDirectionUpdated = await User.findOneAndUpdate(
-        { _id: authUserId, "addressDirections._id": addressDirectionId },
+        { _id: userId, "addressDirections._id": addressDirectionId },
         { $set: { "addressDirection.$": req.body } },
         { new: true },
-      );
+      ).select("addressDirections");
       if (!addressDirectionUpdated) return handleObjectNotFound(res, "Address");
 
       return res.status(200).json(addressDirectionUpdated);
@@ -42,16 +39,17 @@ class AddressDirectionController {
   }
   public async deleteAddressDirection(req: Request, res: Response) {
     try {
-      const { addressDirectionId } = req.params;
-      const authUserId = extractAuthUserId(req);
+      const { userId, addressDirectionId } = req.params;
 
-      const user = await User.findById(authUserId);
+      const user = await User.findById(userId).select("addressDirections");
       if (!user) return handleObjectNotFound(res, "Address");
 
+      if (!user.addressDirections.id(addressDirectionId))
+        return handleObjectNotFound(res, "Address");
       user.addressDirections.id(addressDirectionId)?.deleteOne();
       const userSaved = await user.save();
       if (!userSaved) return handleBadSaved(res);
-      return res.status(201).json(userSaved);
+      return res.status(204).json(userSaved);
     } catch (e) {
       return handleError(res, e);
     }
